add checagem de deleção

+++ sad_app/views.py
@@ -403,48 +403,33 @@ def lista_modelo(request, tabela):
         'lista_modelos': model_list
     }
     
+    if request.method == "POST":
+        list_delete = []
+        for key in request.POST:
+            if key != "csrfmiddlewaretoken":
+                list_delete.append(request.POST[key])
+        
+        try:
+            # deleta os registros   
+            count_delete, model_deletions = model.objects.filter(id__in=list_delete).delete()
+        # exesçao para registros protegidos    
+        except ProtectedError as e: 
+            tabelas_protegidos = []
+            itens_protegidos = []
+            for item in e.protected_objects:
+                if item.__class__.__name__ not in tabelas_protegidos:
+                    tabelas_protegidos.append(item.__class__.__name__)
+                itens_protegidos.append(item)
+            context['itens_protegidos'] = itens_protegidos
+            context['tabelas_protegidos'] = tabelas_protegidos
+            return render(request, 'admin/lista_modelo.html', context)
+        
+        context['count_delete'] = count_delete
+        context['model_deletions'] = model_deletions
+        
     return render(request, 'admin/lista_modelo.html', context)
-
-def deleta_registro(request, tabela):
-    list_delete = []
-    for key in request.POST:
-        if key != "csrfmiddlewaretoken":
-            list_delete.append(request.POST[key])
     
     
-    # procura o model da tabela
-    nome = "sad_app_" + tabela
-    model = None
-    for m in apps.get_models():
-        nome_tabela = m._meta.db_table
-        if nome_tabela == nome:
-            model = m
-    
-    # deleta os registros
-    protegidos = {}
-    try:
-        count_delete, model_deletions = model.objects.filter(id__in=list_delete).delete()
-    except ProtectedError as e:
-        for item in e.protected_objects:
-            if item.__class__.__name__ not in protegidos.keys():
-                protegidos[item.__class__.__name__] = []
-                print(item.__dict__)
-        print(e.__dict__)
-        return HttpResponseRedirect(reverse('lista_modelo', args=[tabela]))
-    
-    # verifica deleções
-    print(model)
-    tabelas_relacionadas = []
-    
-    for tbl in model_deletions:
-        print(tbl)
-    if count_delete != len(list_delete):
-        print("Alguns registros não foram deletados corretamente", count_delete)
-        print("Deletou", model_deletions)
-    else:
-        print("Deletou", count_delete)
-    
-    return HttpResponseRedirect(reverse('lista_modelo', args=[tabela]))
     
     
 def painel_admin(request):
